Updated the CeX wishlist page to be more relevant and also contain the cover image Added game prices to the view game information popup

+++ Models/Gaming.RealmObjects/Game.cs
@@ -13,7 +13,7 @@ namespace Gaming.RealmObjects
         public bool? HasCase { get; set; }
         public bool? HasGame { get; set; }
         public bool? HasManual { get; set; }
-        public int? IgdbId { get; set; }
+        public long? IgdbId { get; set; }
         public string? IgdbUrl { get; set; }
         public bool? IsCollectorsEdition { get; set; }
         public bool? IsCopy { get; set; }
